+++ source/DataTypes.h
@@ -1,0 +1,6 @@
+#ifndef DATA_TYPES_H
+#define DATA_TYPES_H
+
+typedef unsigned char uint8;
+
+#endif
