+++ source/Radio.c
@@ -1,142 +1,10 @@
 #include "cc1110.h"
+#include "Radio.h"
+
 #include <stdint.h>
 #include <string.h>
 #include <stdlib.h>
 #include <time.h>
-
-//Define RFST constant values
-#define SFSTXON             0x00
-#define SCAL                0x01
-#define SRX                 0x02
-#define STX                 0x03
-#define SIDLE               0x04
-
-//Define AX25 Frame Constants
-#define AX25_FRAME_START    0xFE
-#define AX25_CONTROL        0x03
-#define AX25_PROTO_IDENT    0xF0
-#define AX25_FRAME_ID       0x00
-#define AX25_1ST_HEADER_PTR 0xFE
-#define AX25_FRAME_END 	    0xFE
-
-typedef unsigned char uint8;
-
-/********************************************************************************
-*---STRUCT AX25_FRAME---
-*
-* Description:
-*	Comprises one data frame according to the AX25 specification. All data
-*	fields are uint8s. Total length = 251 Bytes.
-*
-*********************************************************************************/	
-typedef struct AX25_Frame
-{
-	//---START---
-	uint8 frame_Start;  //Start of AX25 Frame identifier 0xFE
-	
-	//Header information
-	uint8 dest_Addr[7];         //Callsign of destination station in ASCII
-	uint8 src_Addr[7];          //Callsign of source station in ASCII
-	uint8 control;              //Packet configuration bits fixed to 0x03
-	uint8 proto_Ident;          //Packet configuration bits fixed to 0xF0
-
-	//Information Field
-	uint8 frame_ID;     	    //Packet configuration bits fixed to 0x00
-	uint8 master_Frame_Count;   //Number of current packet being sent
-	uint8 vc_Frame_Count;       //Number of current packet in current virtual channel being sent
-	uint8 first_Header_Pointer; //Not used, hard fixed to 0xFE
-	uint8 data[251];            //Contains message bits provided by mainboard
-	uint8 frame_Status;         //Configures time stamp options
-	uint8 time_Stamp[8];        //Time stamp 	
-
-	uint8 frame_End;            //End of AX25 Frame identifier 0xFE
-	//---END---
-} AX25_Frame;
-
-/********************************************************************************
-*---STRUCT DATA_FRAME---
-*
-* Description:
-*	Comprises one data field to be transferred to or from the communications board
-*	to mainboard.
-*
-*********************************************************************************/	
-typedef struct Data_Frame
-{
-	uint8 data[251];  
-	uint8 dest_Addr[7];
-	uint8 src_Addr[7];
-	uint8 master_Frame_Count;
-	uint8 vc_Frame_Count;
-	uint8 time_Stamp[8];	
-} Data_Frame;
-
-/********************************************************************************
-*---STRUCT AX25_QUEUE---
-*
-* Description:
-*	Queue for buffering AX25_Frame TX and RX transfers between ground and 
-*	the satellite.
-*
-*********************************************************************************/	
-typedef struct AX25_Queue
-{
-	uint8       head;
-	uint8       tail;
-	uint8       size;
-	uint8       capacity;
-	AX25_Frame* AX25_Frame_Array_List;	
-} AX25_Queue;
-
-/********************************************************************************
-*---STRUCT DATA_QUEUE---
-*
-* Description:
-*	Queue for buffering 251-Byte data field transfers to and from mainboard.
-*
-*********************************************************************************/	
-typedef struct Data_Queue
-{
-	uint8       head;
-	uint8       tail;
-	uint8       size;
-	uint8       capacity;
-	Data_Frame* Data_Frame_Array_List;	
-} Data_Queue;
-
-
-//---FUNCTION DECLARTIONS---
-
-//Set-Up
-void        initConfigRegisters 	(void);
-void        initClock           	(void);
-void        initRFStateMach     	(void);
-
-//Frame Manipulation
-Data_Frame* decomm_AX25_Packet  	(struct AX25_Frame *frame);
-
-//AX25 Queue Operation
-AX25_Queue* create_AX25_Queue   	(uint8 capacity);
-uint8       isFullAX25          	(AX25_Queue* queue);
-uint8       isEmptyAX25         	(AX25_Queue* queue);
-uint8       enqueueAX25         	(AX25_Queue* queue, AX25_Frame* frame);
-AX25_Frame* dequeueAX25         	(AX25_Queue* queue);
-
-//Data Frame Queue Operation
-Data_Queue* create_Data_Queue   	(uint8 capacity);
-uint8  	    isFullData          	(Data_Queue* queue);
-uint8       isEmptyData         	(Data_Queue* queue);
-uint8       enqueueData         	(Data_Queue* queue, Data_Frame* data);
-Data_Frame* dequeueData                 (Data_Queue* queue);
-
-//Transmit Operation
-uint8       transmit_AX25_Frame         (AX25_Frame* frame);
-
-//Receive Operation
-AX25_Frame* receive_AX25_Frame          (void);
-
-//Helper Functions For AX25 Frame Construction
-AX25_Frame* construct_AX25_Packet       (Data_Frame* frame);
 
 void main(void)
 {
@@ -237,272 +105,7 @@ void initRFStateMach(void)
 	RFTXRXIF = 0;
 }
 
-/********************************************************************************
-*---FUNCTION---
-* Name: create_AX25_Queue()
-* Description:
-*	Creates and allocates memory for an AX25 RX or TX queue.
-* Parameters:
-*	uint8 capacity - capacity of the queue
-* Returns:
-*	AX25_Queue* - pointer to the create AX25_Queue object
-*********************************************************************************/	
-AX25_Queue* create_AX25_Queue(uint8 capacity)
-{
-	//Allocate space for an AX25_Queue
-	AX25_Queue* queue = (AX25_Queue *) malloc(sizeof(AX25_Queue));
-	
-	//Set the capacity of the queue
-	queue->capacity = capacity;
-	
-	//Set the counters for head and tail and set the size of the queue to 0
-	queue->head = 0;
-	queue->size = 0;
-	queue->tail = capacity - 1;
 
-	//Allocate space for the list of AX25_Frames in the queue
-	queue->AX25_Frame_Array_List = (AX25_Frame *) malloc(sizeof(AX25_Frame) * queue->capacity);	
-
-	return queue;
-}
-
-/********************************************************************************
-*---FUNCTION---
-* Name: isFullAX25()
-* Description:
-*	Returns true if the size of the queue is equal to the capacity of the
-*       queue.
-* Parameters:
-*	AX25_Queue* - pointer to the queue	
-* Returns:
-*	uint8 - boolean value, 1 if queue is full, 0 if queue is not
-*	full	
-*********************************************************************************/	
-uint8 isFullAX25(AX25_Queue* queue)
-{
-	return(queue->size == queue->capacity);
-}
-
-/********************************************************************************
-*---FUNCTION---
-* Name: isEmptyAX25()
-* Description:
-*	Returns true if the size of the queue is equal to 0.
-* Parameters:
-*	AX25_Queue* - pointer to the queue	
-* Returns:
-*	uint8 - boolean value, 1 if queue is empty, 0 if queue is not
-*	empty	
-*********************************************************************************/	
-uint8 isEmptyAX25(AX25_Queue* queue)
-{
-	return(queue->size == 0);
-}
-
-/********************************************************************************
-*---FUNCTION---
-* Name: enqueueAX25()
-* Description:
-*	Enqueues the passed AX25_Frame at the tail of the queue.
-* Parameters:
-*	AX25_Queue* - pointer to the queue
-*	AX25_Frame* - pointer to the frame being enqueued	
-* Returns:
-*	uint8 - boolean value, 1 if enqueue fails, 0 if it succeeds
-*********************************************************************************/	
-uint8 enqueueAX25(AX25_Queue* queue, AX25_Frame* frame)
-{
-	//Check to make sure the queue has available space for an additional frame
-	if(isFullAX25(queue))
-	{
-		return 1;
-	}
-		
-	//Adjust the tail counter for the queue
-	queue->tail = (queue->tail + 1) % queue->capacity;
-
-	//Copy the memory from the frame into the queue
-	memcpy(&(queue->AX25_Frame_Array_List[queue->tail]), frame, sizeof(*frame));
-	
-	//Update the size of the queue
-	queue->size = queue->size + 1;
-
-	//Free the memory of the element since it is now in the queue
-	free(frame);
-	
-	return 0;
-}
-
-/********************************************************************************
-*---FUNCTION---
-* Name: dequeueAX25()
-* Description:
-*	Dequeues the the AX25_Frame at the head of the queue.
-* Parameters:
-*	AX25_Queue* - pointer to the queue
-* Returns:
-*	AX25_Frame* - dequeued frame
-*********************************************************************************/	
-AX25_Frame* dequeueAX25(AX25_Queue* queue)
-{
-	AX25_Frame* frame;
-	
-	//Check to make sure the queue has an available element to be dequeued
-	if(isEmptyAX25(queue))
-	{
-		return NULL;
-	}
-
-	//Allocate memory to store the dequeued frame bits
-	frame = (AX25_Frame *) malloc(sizeof(AX25_Frame));	
-
-	//Copy the frame bits from the element in the queue to the new element
-	memcpy(frame, &(queue->AX25_Frame_Array_List[queue->head]), sizeof(queue->AX25_Frame_Array_List[queue->head]));
-
-	//Set the bits of the element in the queue (the queue is fixed capacity so its memory will not be reclaimed)
-	memset(&(queue->AX25_Frame_Array_List[queue->head]), 0x00, sizeof(AX25_Frame));
-
-	//Update the head counter and size values for the queue
-	queue->head = (queue->head + 1) % queue->capacity;
-	queue->size = queue->size - 1;
-
-	return frame;
-}
-
-/********************************************************************************
-*---FUNCTION---
-* Name: create_Data_Queue()
-* Description:
-*	Creates and allocates memory for a Data RX or TX queue.
-* Parameters:
-*	uint8 capacity - the capacity of the queue
-* Returns:
-*	Data_Queue* - pointer to the created data queue
-*********************************************************************************/	
-Data_Queue* create_Data_Queue(uint8 capacity)
-{
-	//Allocate memory for the new Data_Queue
-	Data_Queue* queue = (Data_Queue *) malloc(sizeof(Data_Queue));
-	
-	//Set the capacity of the queue
-	queue->capacity = capacity;
-	
-	//Set the counter for the head and tail of the queue, set the size of the queue
-	queue->head = 0;
-	queue->size = 0;
-	queue->tail = capacity - 1;
-	
-	//Allocate memory for the list of Data_Frames in the queue
-	queue->Data_Frame_Array_List = (Data_Frame *) malloc(sizeof(Data_Frame) * queue->capacity);	
-
-	return queue;
-
-}
-
-/********************************************************************************
-*---FUNCTION---
-* Name: isFullData()
-* Description:
-*	Returns true if the size of the queue is equal to the capacity of the
-*       queue.
-* Parameters:
-*	Data_Queue* - pointer to the queue	
-* Returns:
-*	uint8 - boolean value, 1 if queue is full, 0 if queue is not
-*	full	
-*********************************************************************************/	
-uint8 isFullData(Data_Queue* queue)
-{
-	return(queue->size == queue->capacity);
-}
-
-/********************************************************************************
-*---FUNCTION---
-* Name: isEmptyData()
-* Description:
-*	Returns true if the size of the queue is equal to 0.
-* Parameters:
-*	Data_Queue* - pointer to the queue	
-* Returns:
-*	uint8 - boolean value, 1 if queue is empty, 0 if queue is not
-*	empty	
-*********************************************************************************/	
-uint8 isEmptyData(Data_Queue* queue)
-{
-	return(queue->size == 0);
-}
-
-/********************************************************************************
-*---FUNCTION---
-* Name: enqueueData()
-* Description:
-*	Enqueues the passed Data_Frame at the tail of the queue.
-* Parameters:
-*	Data_Queue* - pointer to the queue
-*	Data_Frame* - pointer to the frame being enqueued	
-* Returns:
-*	uint8 - boolean value, 1 if enqueue fails, 0 if it succeeds
-*********************************************************************************/	
-uint8 enqueueData(Data_Queue* queue, Data_Frame* data)
-{
-	//Check to make sure the queue has available space to enqueue the new item
-	if(isFullData(queue))
-	{
-		return 1;
-	}
-	
-	//Adjust the position of the tail counter
-	queue->tail = (queue->tail + 1) % queue->capacity;
-
-	//Copy the bits from the Data_Frame into the queue
-	memcpy(&(queue->Data_Frame_Array_List[queue->tail]), data, sizeof(*data));
-
-	//Update the size of the queue
-	queue->size = queue->size + 1;
-
-	//Free the memory for the original Data_Frame
-	free(data);
-	
-	return 0;
-}
-
-/********************************************************************************
-*---FUNCTION---
-* Name: dequeueData()
-* Description:
-*	Dequeues the the Data_Frame at the head of the queue.
-* Parameters:
-*	Data_Queue* - pointer to the queue
-* Returns:
-*	Data_Frame* - dequeued frame
-*********************************************************************************/	
-Data_Frame* dequeueData(Data_Queue* queue)
-{
-	Data_Frame* data;
-	
-	//Check to make sure the queue has at least element that can be dequeued
-	if(isEmptyData(queue))
-	{
-		return NULL;
-	}
-
-	//Allocate space to return the dequeued Data_Frame
-	data = (Data_Frame *) malloc(sizeof(Data_Frame));
-
-	//Copy over the bits from the element in the queue to new Data_Frame	
-	memcpy(data, &(queue->Data_Frame_Array_List[queue->head]), sizeof(queue->Data_Frame_Array_List[queue->head]));
-	
-	//Set the bits in the queue (since the queue is fixed capacity memory will not be freed, this "clears" the memory)
-	memset(&(queue->Data_Frame_Array_List[queue->head]), 0x00, sizeof(Data_Frame));
-	
-	//Adjust the head pointer of the queue
-	queue->head = (queue->head + 1) % queue->capacity;
-	
-	//Update the size of the queue
-	queue->size = queue->size - 1;
-
-	return data;
-}
 
 /********************************************************************************
 *---FUNCTION---
